feat: migration of spring sdk test modules
Introduces spring sdk test modules from
former community sdk.
https://github.com/camunda-community-hub/spring-zeebe/tree/main/test
